@@ -1,6 +1,6 @@
-///Get Navigator allows you to navigate routes, open snackbars,
+///Jet Navigator allows you to navigate routes, open snackbars,
 ///dialogs and bottomsheets easily, and without the need for context.
 library;
 
-export 'get_core/get_core.dart';
-export 'get_navigation/get_navigation.dart';
+export 'jet_core/jet_core.dart';
+export 'jet_navigation/jet_navigation.dart';
