@@ -1,6 +1,6 @@
-/// Get State Manager is a light, modern and powerful state manager to Flutter
+/// Jet State Manager is a light, modern and powerful state manager to Flutter
 library;
 
-export 'get_core/get_core.dart';
-export 'get_rx/get_rx.dart';
-export 'get_state_manager/get_state_manager.dart';
+export 'jet_core/jet_core.dart';
+export 'jet_rx/jet_rx.dart';
+export 'jet_state_manager/jet_state_manager.dart';
