@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:get/get.dart';
+import 'package:jet/jet.dart';
 
-import 'get_main_test.dart';
+import 'jet_main_test.dart';
 
 class RedirectMiddleware extends GetMiddleware {
   @override
@@ -33,111 +33,111 @@ class RedirectBypassMiddleware extends GetMiddleware {
 
 void main() {
   tearDown(() {
-    Get.reset();
+    Jet.reset();
   });
 
   testWidgets("Middleware should redirect to second screen", (tester) async {
     // Test setup
     await tester.pumpWidget(
-      GetMaterialApp(
+      JetMaterialApp(
         initialRoute: '/',
-        getPages: [
-          GetPage(name: '/', page: () => const Home()),
-          GetPage(
+        jetPages: [
+          JetPage(name: '/', page: () => const Home()),
+          JetPage(
             name: '/first',
             page: () => const FirstScreen(),
             middlewares: [RedirectMiddleware()],
           ),
-          GetPage(name: '/second', page: () => const SecondScreen()),
-          GetPage(name: '/third', page: () => const ThirdScreen()),
+          JetPage(name: '/second', page: () => const SecondScreen()),
+          JetPage(name: '/third', page: () => const ThirdScreen()),
         ],
       ),
     );
 
     // Act
-    Get.toNamed('/first');
+    Jet.toNamed('/first');
     await tester.pumpAndSettle();
 
     // Assert
     expect(find.byType(SecondScreen), findsOneWidget);
     expect(find.byType(FirstScreen), findsNothing);
-    expect(Get.currentRoute, '/second');
+    expect(Jet.currentRoute, '/second');
   });
 
   testWidgets("Middleware should stop navigation", (tester) async {
     // Test setup
     await tester.pumpWidget(
-      GetMaterialApp(
+      JetMaterialApp(
         initialRoute: '/',
-        getPages: [
-          GetPage(name: '/', page: () => const Home()),
-          GetPage(
+        jetPages: [
+          JetPage(name: '/', page: () => const Home()),
+          JetPage(
             name: '/first',
             page: () => const FirstScreen(),
             middlewares: [RedirectMiddlewareNull()],
           ),
-          GetPage(name: '/second', page: () => const SecondScreen()),
-          GetPage(name: '/third', page: () => const ThirdScreen()),
+          JetPage(name: '/second', page: () => const SecondScreen()),
+          JetPage(name: '/third', page: () => const ThirdScreen()),
         ],
       ),
     );
 
     // Act
     await tester.pumpAndSettle();
-    Get.toNamed('/first');
+    Jet.toNamed('/first');
     await tester.pumpAndSettle();
 
     // Assert
     expect(find.byType(Home), findsOneWidget);
     expect(find.byType(FirstScreen), findsNothing);
-    expect(Get.currentRoute, '/');
+    expect(Jet.currentRoute, '/');
   });
 
   testWidgets("Middleware should be bypassed", (tester) async {
     // Test setup
     await tester.pumpWidget(
-      GetMaterialApp(
+      JetMaterialApp(
         initialRoute: '/',
-        getPages: [
-          GetPage(name: '/', page: () => const Home()),
-          GetPage(
+        jetPages: [
+          JetPage(name: '/', page: () => const Home()),
+          JetPage(
             name: '/first',
             page: () => const FirstScreen(),
             middlewares: [RedirectBypassMiddleware()],
           ),
-          GetPage(name: '/second', page: () => const SecondScreen()),
-          GetPage(name: '/third', page: () => const ThirdScreen()),
+          JetPage(name: '/second', page: () => const SecondScreen()),
+          JetPage(name: '/third', page: () => const ThirdScreen()),
         ],
       ),
     );
 
     // Act
     await tester.pumpAndSettle();
-    Get.toNamed('/first');
+    Jet.toNamed('/first');
     await tester.pumpAndSettle();
 
     // Assert
     expect(find.byType(FirstScreen), findsOneWidget);
     expect(find.byType(SecondScreen), findsNothing);
     expect(find.byType(Home), findsNothing);
-    expect(Get.currentRoute, '/first');
+    expect(Jet.currentRoute, '/first');
   });
 
   testWidgets("Middleware should redirect twice", (tester) async {
     // Test setup
     await tester.pumpWidget(
-      GetMaterialApp(
+      JetMaterialApp(
         initialRoute: '/',
-        getPages: [
-          GetPage(name: '/', page: () => const Home()),
-          GetPage(
+        jetPages: [
+          JetPage(name: '/', page: () => const Home()),
+          JetPage(
             name: '/first',
             page: () => const FirstScreen(),
             middlewares: [RedirectMiddleware()],
           ),
-          GetPage(name: '/second', page: () => const SecondScreen()),
-          GetPage(name: '/third', page: () => const ThirdScreen()),
-          GetPage(
+          JetPage(name: '/second', page: () => const SecondScreen()),
+          JetPage(name: '/third', page: () => const ThirdScreen()),
+          JetPage(
             name: '/fourth',
             page: () => const FourthScreen(),
             middlewares: [Redirect2Middleware()],
@@ -147,47 +147,47 @@ void main() {
     );
 
     // Act
-    Get.toNamed('/fourth');
+    Jet.toNamed('/fourth');
     await tester.pumpAndSettle();
 
     // Assert
     expect(find.byType(SecondScreen), findsOneWidget);
     expect(find.byType(FirstScreen), findsNothing);
-    expect(Get.currentRoute, '/second');
+    expect(Jet.currentRoute, '/second');
   });
 
   testWidgets("Navigation history should be correct after redirects",
       (tester) async {
     // Test setup
     await tester.pumpWidget(
-      GetMaterialApp(
+      JetMaterialApp(
         initialRoute: '/',
-        getPages: [
-          GetPage(name: '/', page: () => const Home()),
-          GetPage(
+        jetPages: [
+          JetPage(name: '/', page: () => const Home()),
+          JetPage(
             name: '/first',
             page: () => const FirstScreen(),
             middlewares: [RedirectMiddleware()],
           ),
-          GetPage(name: '/second', page: () => const SecondScreen()),
+          JetPage(name: '/second', page: () => const SecondScreen()),
         ],
       ),
     );
 
     // Act
-    Get.toNamed('/first');
+    Jet.toNamed('/first');
     await tester.pumpAndSettle();
 
     // Assert
-    expect(Get.currentRoute, '/second');
-    expect(Get.previousRoute, '/');
+    expect(Jet.currentRoute, '/second');
+    expect(Jet.previousRoute, '/');
 
     // Act: go back
-    Get.back();
+    Jet.back();
     await tester.pumpAndSettle();
 
     // Assert
     expect(find.byType(Home), findsOneWidget);
-    expect(Get.currentRoute, '/');
+    expect(Jet.currentRoute, '/');
   });
 }
